+++ src/types/payroll.ts
@@ -1,3 +1,4 @@
 export type PayrollData = {
   name: string;
+  familySituation: "soltero" | "casado" | "otros";
 };